pylib: Move some classes away from client/__init__.py
Makes those files somewhat cleaner, and avoids some issues with imports
for type hinting.
