+++ .bumpversion.cfg
@@ -1,7 +1,7 @@
 [bumpversion]
 current_version = 3.11.0
 
-[bumpversion:file:pylib/openrazer/client/__init__.py]
+[bumpversion:file:pylib/openrazer/client/device_manager.py]
 
 [bumpversion:file:daemon/openrazer_daemon/daemon.py]
